Remove repurposed commands from ColumnsSheet
Previously, ~!@#$%_-+ commands to set column types, widths, and keys were
redone on the ColumnsSheet to operate on the "rows" (columns on the
source sheet).  This was clever but ultimately confusing and difficult
when the actual ColumnsSheet columns needed tweaking.  The individual
commands have been removed and should now act consistently across
all sheets.  The original workflow desire--to easily modify
source column attributes from its ColumnsSheet--is satisfied
with 'g' versions of the original commands that operate on all
selected "rows", which are mostly unspecified in the global context.
Exception is `g_`, which is still used on other sheets to reset
the width of all visible columns, but on the ColumnsSheet is repurposed
to act like the other 'g' commands on that sheet.