[#250] Use value_counts() for PandasSheetFreqTable
An initial hacky implementation for using value_counts()
to generate frequency table for pandas, which is vectorized
and hence significantly faster than the approach in
PivotTable. Added support for [Enter] > sourcerows.
Toggling/selection support is still broken here because
id(row) will not work with pandas.

Also tried to do some small cleanup with dtype
handling in PandasSheet.

This commit includes a lot of experimental and debugging
code which can be removed.

todo: support toggle in PandasSheet/PandasSheetFreqTable