[undo-] redo with cmd on first row did not move cursor; fixes #339
The cmdlog was storing the integer row index as the rowname.
When saved, this index is converted to a string in the .vd file.
On replay, moveToReplayContext moves the row cursor to the given row.
It uses `if r.row` to detect if a row was specified.
When the row is `0` (int, not str), this check is false.
Redo replays the undone command, and since the row context hasn't been stringified to a .vd, it doesn't detect the given row if it is 0.
So redoing an undone command that was executed from the first row, will redo it in an improper context.

This fix sets the row/col context as strings, even when they are numeric indexes.