hub: First implementation of The Hub
The Flask app that will serve the main website as an SPA.
Here the SPA has been built using the default HTML/JS from gotty, it's
been somewhat cleaned and refactored.

The Hub will also be able to schedule dedicated vd containers and proxy
websockets to those containers. This commit does not include that yet.