frontend: Mithril, MDC framework, auth, layout
First implementation of SPA using Mithril. This sets out the basic
scaffolding and design for managing the frontend of the app.

There is a fake /api/auth endpoint to at least glue the user flow
together.

NB. I tried to get JSX working for TSX files, but it doesn't like
Mithril :/