[tests] Add failing tests demonstrating a tricky bug in namedlist
SequenceSheet rows can not have columns named "length" or "_fields" or
"__name__" due to conflicts with existing attributes on the namedlist
implementation.  This affects formats like csv and tsv but not jsonl,
since the latter does not use SequenceSheet.

The "length" bug was reported in #344 and fixed back in Sept 2019, but
since then the bug was apparently re-introduced.  I did not bisect to
find the re-introduction.