[tests] Fail more clearly if a test is missing a golden output
The previous error message was:

    no function to save as type *
    fail: no function to save as type *

because the glob pattern failed to match and was passed as-is as the
golden file.

The new error message is:

    ./dev/test.sh: line 25: no match: tests/golden/errors.*

This only applies to tests not marked "nosave".

Builds upon the intent of "[tests] expand test.sh to work with non-tsv
goldens" (56723a9b), which switched to determining each test's output
files by the existing files within tests/golden/.  Some tests are
missing golden outputs, but before that commit, they still had their
output files saved to tsv.