[cmdlog risky] addUndo directly modifies cmdlog row if activeCommand has already been logged
WARNING: risky commit;

It is still possible to find race conditions if
the user presses commands fast enough, however, this will allow some grace for complex Undos.
If we want to further reduce race conditions, we can track sheet threads
and block more than one modification, but that is a much bigger change.

Closes #700, #668