[undo-] to avoid race condition, do not track undos for CommandLogs
Related to 16e0a606aa31

Essentially, when the first CommandLog is created, it needs to set
Undos, and there are not any cmdlog.rows for it to pull from in the
pre-command times.

This case is usually handled by if not vd.sheet, except when you replay,
a .vd (CommandLog) is loaded in the pre-command times, when the main
CommandLog is still not ready to stage undos

This change means that we do not have undo functionalities on any
CommandLog sheets. It was decided that this is better than having
unpredictable undos, but this should be addressed in the future.