[WIP bugfix]: in splitpane, mouse actions should be registered where clicked
Action: Open two sheets, do splitpane, attempt to drag column on
non-active sheet.

Expected: column is dragged on non-active sheet.

Actual: the column in that viscinity on the active sheet is dragged.

What is currently working: both panes can be successfully dragged, when
they are active. The nature of the bug is that an active sheet's columns are
being un-intuitively dragged, even if the user is dragging on the opposite
panel.