Fixed mysql loader duplicating tables for each database
The current implementation duplicates the tables of the database
you selected for all available databases because of a wrong join.
So if I have 2 databases A (containing table a1 and a2) and B (containing
table b1 and b2) and I open visidata on database A, i will see:
a1, a2, a1, a2
This makes no sense at all.