[mysql-loader] asynchronous data fetch for mysql loader
Unless SSCursor (see https://stackoverflow.com/questions/9395793/mysqldb-and-big-queries-using-cursoruseresultmixin)
is used the mysql driver will first fetch all data and only when all data is
fetched it will pass it to python.
This causes you to wait for all data to be transfered before
it is displayed in visidata. With this fix you see data
immediately and the rest of the data loads in the background.