[airtable] Wrap API calls in Progress
For loading, Airtable doesn't tell us how many total pages there
are, but we can at least increase the progress with each additional
page. For saving, unfortunately we can't reach into the library
which is thankfully dealing with splitting the batches and respecting
the rate limit, but we can at least indicate which stage of saving
is being performed.