[color-] determine color availability with init_pair
with non-256-color terminals (or non-256 $TERM in the case of
tmux/screen), higher color numbers were still accepted but always
returned the default color, when they should be rejected and fall to the
next color as a fallback (like "218 cyan").  This changed with
3a91664517394a5d14414992dd5ccb2fb95c8b34.

instead of using curses.COLORS (perhaps forced to 256 with
options.force_256_colors) to determine whether a color is valid, try
init_pair with it on a black bg.  if that fails, mark the color as
unavailable and give the next color a chance.