[join-] fix multiple extend
The rowdef on joined sheets was a list of references to the rows on the source sheets.
Columns on joined sheets got the row ref by numeric index.

Columns on 'extend' joined sheets, however, depended on rowsBySheetKey, a dict saved on the joined sheet.
With multiple extend joins, the rowsBySheetKey on intermediate sheets was unavailable to columns on the final sheet.

Now a ref to rowsBySheetKey is stored on the columns themselves, and
rowdef for joined sheets is now a dict of row references, by sheet object.

This fixes the issue, and is more robust in general.

Thanks to @cwarden for reporting.