[undo-] call undofuncs in reverse order #1133
Both setValues and setValue call addUndo, and setValues calls setValue.
So multiple undofuncs are added for some commands, like delete-cell.
Calling these undofuncs as implemented and in the order they are added, causes some internal confusion leading to this warning message, but actually nothing is incorrect and the results are the same.
Ideally, commands would not add more undofuncs than necessary (it can waste memory), but that's trickier than this fix, which just calls undofuncs in reverse LIFO order, so that subsequent undofuncs undo previous themselves without concern.