Update data.fixer.io using apilayer with requests
The original currency conversion API data.fixer.io
redirects already to apilayer.com and I couldn't find
a way to create API keys for the former data.fixer.io API,
so I updated the url and passing of parameters.
Especially the api key is passed now as header instead as a url param
what's certainly reasonable for security.

Another problem was that apilayer blocks urllib requests as bot requests
(for at least for the free account) and following
https://www.pythonpool.com/urllib-error-httperror-http-error-403-forbidden/
I changed the method to requests library as also this is officially
supported by apilayer as can be seen in:
https://apilayer.com/marketplace/fixer-api?live_demo=show

It would be probably a good idea to change to requests in general,
but I'm not sure how to test it accordingly, that's why a parameter to
the urlcache function was added, so you can switch between urllib and
requests.

Still contains for convenience logging messages that will be removed
next.