Remove debug logging
Wasn't sure how to properly debug it
(import pdb; pdb.set_trace() collided with the curses frontend),
so wrote a lot of logging debug outputs.
Won't be needed any longer, but in case they would,
it might be nice to have them in an own commit,
so did not squeeze them with the other changes.