Allow currency caching days to be configurable
Was easier for development than removing caching again,
and I think for at least no caching (days=0) or some caching (days=1)
should be part when working with financial data.

For a rough overview in an exploratory data analysis,
fast cached data should be enough,
but it should certainly be switchable off
in cases where we'd like to have the up-to-date currency rates.