Adds unguard-sheet command
IMHO a guarding of a sheet should be reverseable,
so added this simple command.

Not sure if we should show a status,
when guarded it's obviously we show,
but I decided against showing a status,
as unguarded is the normal behaviour and see no reason to differentiate
in the status between never guarded or once guarded and then unguarded.