Fix clipboard on Windows
Windows doesn't allow files to be opened for reading if they are already
open.

For syscopyValues:
Previously, we created and opened a temporary file, then opened it
two more times while it was already open.

However, NamedTemporaryFile creates AND opens a file, so now we open it
with the read-write mode 'w+' and use seek(0) so we can reuse a single
open file handle for everything.

For syscopyCells_async:
We can't reuse the open file handle here without refactoring
vd.saveSheets, as it relies on being passed a file path with an
extension to determine which saver to use. So, we take the pragmatic
option, and just close the file as soon as we create it to make Windows
happy, then open and close two more times, once to save the data, and
once to read it.

Fixes #1431