Fix clipboard on Windows (pipe to clip cmd stdin)
Windows doesn't allow files to be opened for reading if they are already
open.

For syscopyValues:
Previously, we created and opened a temporary file, then opened it
two more times while it was already open.

Now we just pipe the data directly to the clipboard command's stdin.

For syscopyCells_async:
We can't reuse the open file handle here without refactoring
vd.saveSheets, as it relies on being passed a file path with an
extension to determine which saver to use. So, we take the pragmatic
option, and just close the file as soon as we create it to make Windows
happy, then open and close two more times, once to save the data, and
once to read it.

Fixes #1431