[DirSheet-] change open-row to open-file-row #1527
Command variations on subsheets must have different longnames than the
base command they override.  This is apparently a bug in the command
settings, but also the user may want to execute the original command
anyway--in this case, opening a sheet of the file metadata, instead of
opening the file itself.