[extensible-] do not copy lazy_property attributes #1780
This was a subtle and terrible bug.

BaseSheet.__copy__ calls ret.__dict__.update(self.__dict__), so all
attributes are copied over; and they will be references (since it is a
shallow copy).  This is avoided for attributes created with
Extensible.init(copy=False), since they are re-created with the init
function after this update in the copy, and so they are set to fresh
values.  All other attributes, including those created by
@Extensible.lazy_property, won't be re-set.

There are some other lazy_property attributes that would be affected by
this, like the sheet-specific optionsSheet, and _deferredAdds/Mods/Dels.
But the lazy_property with the most obvious consequences, was
cmdlog_sheet, the cmdlog for this sheet only.  All copied sheets had the
same cmdlog_sheet as the sheet they were copied from, even though they
should have their own.  So commands for all dup'ed sheets would
erroneously share the cmdlog, and undo (which uses this cmdlog_sheet to
find the last modification to undo) would appear to be a more global
undo, as an undo on any of these copied sheets would undo the last
modification that happened to any of them.

This change makes lazy_property create its persisted attributes using .init(copy=False), so they won't be copied.
There is no way to make a @lazy_property share its result across copied sheets; this would have to be done manually.