[xml-] silence FutureWarning from lxml
lxml elements, in a conditional, behave like a list of their children. An element with children evalutes to True, one with no children evaluates to False.
Because the behavior is surprising to people, lxml now gives FutureWarnings when the truth value of an Element is checked via "if lxml_elem:". See
https://lxml.de/tutorial.html#elements-are-lists