[sort-] do not modify Sheet._ordering #2254
Every sheet instance needs its own _ordering list, which orderBy modifies.
But _ordering is defined at the class level so it can be easily specified per class.
Now we make a copy of the class _ordering when a new sheet is instantiated, so that orderBy doesn't modify the class list.