[canvas-] round consistently for point/label plots #2487
For GraphSheet, the calculation of the y-coordinate for points
was slightly inconsistent with the y-coordinate for labels. The
points calculation attempted to inline scaleY(), but it performed
rounding in a slightly different way. As a result, the label for
y=3.0 could sometimes be one square higher than the points with
y=3.0. However, the vertical alignment of points was always
consistent relative to other points.

Due to similar inconsistencies, add_x/y_axis_label was sometimes
placing x/y labels one square away from where a corresponding data
point would be drawn.