[cliptext-] rework _clipstr to avoid backtracking
It is incorrect to shrink a string by trimming characters from the
end of the string, because that can cut into a grapheme cluster.
Avoiding such cuts improves the handling of grapheme clusters. By
itself, the change is not enough to fix how _clipstr breaks up
grapheme clusters, but it makes a future fix simpler.