[sheets-] drop offscreen col from _visibleColLayout
When the right side of the screen contained a column separator
or the last character of a column, _visibleColLayout would
incorrectly include the offscreen column to the right.

The inclusion was mostly harmless, as most retrievals from
_visibleColLayout check the x coordinate of the column for
validity. But allAggregators does not check it, so aggregators
for that offscreen column were shown when they should not
have been visible.

And if disp_column_sep was '', two such offscreen columns were
included, instead of just one.