[cmdlog-] make vdj shebang work on macOS
This is similar to, but should work better than the now-reverted
https://github.com/saulpw/visidata/commit/226d428c8f91f32d4a3dede657d0bc5dfdd24d78

Currently, I get this error if I try to execute a vdj file on my Mac:

```console
$ chmod +x test.vdj
$ ./test.vdj
exec: Failed to execute process './test.vdj': The file specified the interpreter 'vd -p', which is not an executable command.
```

With this change, it works.

-----------------

Some references for the `-S` option:

- `man env`
- https://docs.astral.sh/uv/guides/scripts/#using-a-shebang-to-create-an-executable-file
is another example of using it.