make asynccache/addcol-shell thread-safe, and runnable in macros and replay (#2826)
* [threads-] asynccache/addcol-shell: fix replay/race condition

The straightforward way to assign to d[k] is not thread-safe:
d[k] = vd.execAsync(_func, k, *args, **kwargs)
because events can happen in this order:
1) _func finishes, assigning func()'s return value to d[k], then
2) execAsync's return value is assigned to d[k], overwriting it.

The same failure happens when replay mode temporarily uses
execSync in place of every execAsync call. Using setdefault() is
safe, assigning to d[k] only if _func hasn't yet done so.
The setdefault() technique is commonly said to be atomic on
CPython, for a standard dictionary with no custom methods.

* [shell-] show error for bad $colname in addcol-shell